edit Pipeline

+++ Jenkinsfile
@@ -33,32 +33,34 @@ pipeline {
 				'''
 				script {
                     try{
-						old_container_id = sh "docker --host $DOCKER_HOST ps -a | grep ilyatrof/ytdl-flask | awk '{ print \$1 }'"
-                        sh '''
-						echo Try to kill cootainer $old_container_id
-                        docker --host $DOCKER_HOST kill $old_container_id
-						'''
+						old_container_id_list = sh "docker --host $DOCKER_HOST ps -a | grep ilyatrof/ytdl-flask | awk '{ print \$1 }'"
+                        for(old_container_id in old_container_id_list){
+							if (old_container_id) {
+								sh '''
+								echo Kill cootainer $old_container_id
+								docker --host $DOCKER_HOST kill $old_container_id
+								'''
+							}
+						}
                     }catch (err) {
                         sh 'echo Kill older cootainers ERROR'
-
                     }
                     try{
 						old_images_id_list = sh "docker --host $DOCKER_HOST images | grep ilyatrof/ytdl-flask | awk '{ print \$3 }'"
-                        for(item in old_images_id_list){
-							println item
+                        for(old_images_id in old_images_id_list){
+							if (old_images_id) {
+							sh '''
+							echo Remove image $old_images_id
+							docker --host $DOCKER_HOST rmi -f $old_images_id
+							'''
+							}
 						}
-						sh 'echo Try to remove older image'
-						//docker images | grep ilyatrof/ytdl-flask | awk '{ print \$3 }'
-                        //sh 'docker --host $DOCKER_HOST rmi -f ilyatrof/ytdl-flask:v${BUILD_NUMBER-1}'
-
                     }catch (err) {
                         sh 'echo Remove older image ERROR'
-
                     }
                 }
                 sh 'echo Deploy new container'
                 sh 'docker --host $DOCKER_HOST run --rm --name ytdl-flask-app -d -p 5000:5000 ilyatrof/ytdl-flask:v${BUILD_NUMBER}'
-					
             }
         }
 		stage('Clearing') {
